+++ src/sll/api/serial.c
@@ -74,15 +74,8 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_decode_object(sll_file_t* f)
 			}
 		case SLL_OBJECT_TYPE_STRING:
 			{
-				sll_object_t* o=sll_create_object(SLL_OBJECT_TYPE_STRING);
 				sll_string_t str;
-				if (!sll_decode_string(f,&str)){
-					sll_string_create(0,&(o->dt.s));
-				}
-				else{
-					o->dt.s=str;
-				}
-				return o;
+				return (sll_decode_string(f,&str)?sll_string_to_object_nocopy(&str):sll_string_to_object(NULL));
 			}
 		case SLL_OBJECT_TYPE_ARRAY:
 		case SLL_OBJECT_TYPE_MAP_KEYS:
@@ -390,9 +383,7 @@ __API_FUNC(serial_decode_string){
 	if (!sll_decode_string(sll_file_from_handle(a),&str)){
 		return SLL_ACQUIRE_STATIC_INT(0);
 	}
-	sll_object_t* o=sll_create_object(SLL_OBJECT_TYPE_STRING);
-	o->dt.s=str;
-	return o;
+	return sll_string_to_object_nocopy(&str);
 }