+++ src/sll/api/vm.c
@@ -18,18 +18,11 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_instruction_to_location(sll_
 	sll_string_index_t fp_i;
 	sll_string_index_t fn_i;
 	sll_file_offset_t ln=sll_get_location(sll_current_runtime_data->a_dt,ii,&fp_i,&fn_i);
-	sll_object_t* fn=sll_create_object(SLL_OBJECT_TYPE_STRING);
-	if (fn_i==SLL_MAX_STRING_INDEX){
-		sll_string_from_pointer(SLL_CHAR("@code"),&(fn->dt.s));
-	}
-	else{
-		sll_string_clone(sll_current_runtime_data->a_dt->st.dt+fn_i,&(fn->dt.s));
-	}
 	sll_object_t* o=sll_create_object(SLL_OBJECT_TYPE_ARRAY);
 	sll_array_create(3,&(o->dt.a));
 	o->dt.a.v[0]=sll_string_to_object(sll_current_runtime_data->a_dt->st.dt+fp_i);
 	o->dt.a.v[1]=sll_int_to_object(ln);
-	o->dt.a.v[2]=fn;
+	o->dt.a.v[2]=(fn_i==SLL_MAX_STRING_INDEX?sll_string_to_object_pointer(SLL_CHAR("@code"),SLL_MAX_STRING_LENGTH):sll_string_to_object(sll_current_runtime_data->a_dt->st.dt+fn_i));
 	return o;
 }